Rename torrentcheker session.py
