+++ src/tribler-core/tribler_core/tests/tools/tracker/udp_tracker.py
@@ -2,7 +2,7 @@ import random
 import struct
 from asyncio import DatagramProtocol, get_event_loop
 
-from tribler_core.modules.torrent_checker.session import MAX_INT32
+from tribler_core.modules.torrent_checker.torrentchecker_session import MAX_INT32
 from tribler_core.tests.tools.tracker.tracker_info import TrackerInfo
 
 UDP_TRACKER_INIT_CONNECTION_ID = 0x41727101980
